% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom.R
\name{domAbundance}
\alias{domAbundance}
\title{Domain Abundance for each sample}
\usage{
domAbundance(fastas, pfamA, cut = "ga", cpus = 1L)
}
\arguments{
\item{fastas}{A \code{character} vector giving the file names of the amino 
acid fasta files.}

\item{pfamA}{\code{character} The path to the Pfam-A.hmm file}

\item{cut}{Parameter controlling model-specific thresholding. Can be ethier
\code{"ga"} or \code{"tc"}. See HMMER 3.1b2 manual for more information.}

\item{cpus}{\code{integer} The number of cpus to use.}

\item{distMethod}{Dissimilarity index, partial match to "manhattan", 
"euclidean", "canberra", "bray" (DEFAULT), "kulczynski", "jaccard", "gower", 
"altGower", "morisita", "horn", "mountford", "raup" , "binomial", "chao", 
"cao" or "mahalanobis". See \link[vegan]{vegdist} for more information.}
}
\value{
A \code{list} object. The first element is the abundance matrix and
the second is the \code{dist} matrix.
}
\description{
Compute Pfam-A domain content dissimilarity between two or more
proteomes.
}
