% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{hmmStat}
\alias{hmmStat}
\title{Compute Hmm Stats}
\usage{
hmmStat(bin, hmmfile)
}
\arguments{
\item{bin}{The location of \code{hmmstat} binary.}

\item{hmmfile}{\code{character} The path to the hmm file.}
}
\value{
The file name of the temporary file where stats were written.
}
\description{
Compute hmm stats.
}
